requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
openai>=1.3.0
anthropic>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pytest>=7.4.0
gymnasium>=0.29.0
stable-baselines3>=2.1.0
unreal-engine-python>=5.1.0
open3d>=0.17.0
scipy>=1.11.0
matplotlib>=3.8.0
wandb>=0.15.0
tqdm>=4.66.0
